+++ nextstep-di/src/test/java/nextstep/di/factory/ConfigurationBeanScannerTest.java
@@ -1,11 +1,6 @@
 package nextstep.di.factory;
 
-import com.google.common.collect.Sets;
 import nextstep.di.factory.beandefinition.BeanDefinition;
-import nextstep.di.factory.example.api.MyJdbcTemplate;
-import nextstep.di.factory.example.repository.JdbcUserRepository;
-import nextstep.di.factory.scanner.ClasspathBeanScanner;
-import nextstep.di.factory.scanner.ConfigurationBeanScanner;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -15,20 +10,16 @@ import java.util.Set;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-class ConfigurationBeanScannerTest {
-
-    private ConfigurationBeanScanner cbs;
-    private ClasspathBeanScanner cbds;
+class ConfigurationBeanScannerTest extends AbstractBeanScannerTest {
 
     @BeforeEach
     void setUp() {
-        cbs = new ConfigurationBeanScanner("nextstep.di.factory.example.");
-        cbds = new ClasspathBeanScanner("nextstep.di.factory.example.");
+        initConfigurationpathBeanScanner();
     }
 
     @Test
     public void register_simple() {
-        Set<BeanDefinition> beanDefinitions = cbs.scan();
+        Set<BeanDefinition> beanDefinitions = cbds.scan();
         BeanFactory beanFactory = new BeanFactory(beanDefinitions);
         beanFactory.initialize();
 
@@ -37,25 +28,8 @@ class ConfigurationBeanScannerTest {
 
     @Test
     public void scanTest() {
-        Set<BeanDefinition> beanDefinitions = cbs.scan();
+        Set<BeanDefinition> beanDefinitions = cbds.scan();
 
         assertThat(beanDefinitions).isNotNull();
     }
-
-    @Test
-    public void register_classpathBeanScanner_통합() {
-        Sets.SetView<BeanDefinition> beanDefinitions = Sets.union(cbs.scan(), cbds.scan());
-        BeanFactory beanFactory = new BeanFactory(beanDefinitions);
-        beanFactory.initialize();
-
-        assertNotNull(beanFactory.getBean(DataSource.class));
-
-        JdbcUserRepository userRepository = beanFactory.getBean(JdbcUserRepository.class);
-        assertNotNull(userRepository);
-
-        MyJdbcTemplate jdbcTemplate = beanFactory.getBean(MyJdbcTemplate.class);
-        assertNotNull(jdbcTemplate);
-        assertNotNull(jdbcTemplate.getDataSource());
-    }
-
 }
